Added wrapping when going North
This one has been left again purposefully on this state (24
failures). The reason is that to wrap I needed to add the concept of
grid limit. Initally you put it on the rover because seems easy to add
it there. But then initializ-rover calls need to change all over the
place and the fact that is grid limit is clearly not a concept that a
rover should know about. In fact x and y could be argued that are not
really a concern of the rover. Furthermore, are the exact directions a
concern of the rover itself? limits, changing position and direction
are the concern of a different entity to which rover could ask
questions.

So the next step will be to refactor the code to represent this information.