Turned around rover world relationship
Basically inverted what I did on the previous commit. The rover now
knows about the world in which is located. Two consequences, we can
more logically ask the rover to turn, and we will be able to change a
bit the behaviour so the rover will ask the world what happens when it
turns right or left.