Refactoring concepts
The easy one, eliminated unneded builder. 1 was a funky magic number,
and now the speed concept has been introduced.

Also, move-forwards and mobe-backwards now represent moves of the
rover in the world. The rover says wants to move forwards/backwards at
a certain speed, and its position on the world changes because of it.